stem-connect-backend-template/requirements.txt
Flask
Flask-CORS
pyserial
requests
